BOAC-2996, reusable appointment dropdown component for waitlist, student profile
